refactor: create abstract transfer base scenario runner
