+++ __test__/integration/debt_token_api/debt_token_scenario_runner.ts
@@ -35,7 +35,7 @@ export class DebtTokenScenarioRunner {
     private ownerOfScenarioRunner: OwnerOfScenarioRunner;
     private existsScenarioRunner: ExistsScenarioRunner;
     private transferFromScenarioRunner: TransferFromScenarioRunner;
-    private transferScenarioRunner: TransferFromScenarioRunner;
+    private transferScenarioRunner: TransferScenarioRunner;
     private approveScenarioRunner: ApproveScenarioRunner;
     private getApprovedScenarioRunner: GetApprovedScenarioRunner;
